Computer_Vision/detect_space_full/parking-space-counter/requirements.txt
opencv-python==4.6.0.66
scikit-learn==1.1.3
pandas==1.5.1
Pillow==9.3.0
scikit-image
matplotlib==3.6.2
